backend/requirements.txt
alembic==1.13.3
asgiref==3.7.2
asttokens==2.4.1
attrs==24.2.0
autopep8==2.0.4
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.1
django-bootstrap==0.2.4
django-bootstrap4==24.1
django-crispy-forms==2.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Babel==2.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
greenlet==3.1.1
idna==3.10
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mypy==1.8.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.1.1
nest-asyncio==1.5.8
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parameterized==0.9.0
parso==0.8.3
passlib==1.7.4
pbr==6.0.0
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.11.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
ufw==0.36.1
ujson==5.9.0
urllib3==2.2.3
uvloop==0.19.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.4
